+++ tools/swift-inspect/Sources/libswiftInspect/RemoteMirror+Extensions.swift
@@ -13,9 +13,9 @@
 import SwiftRemoteMirror
 
 extension swift_metadata_allocation_t {
-  internal var tag: swift_metadata_allocation_tag_t { return self.Tag }
-  internal var ptr: swift_reflection_ptr_t { return self.Ptr }
-  internal var size: Int { return Int(self.Size) }
+  public var tag: swift_metadata_allocation_tag_t { return self.Tag }
+  public var ptr: swift_reflection_ptr_t { return self.Ptr }
+  public var size: Int { return Int(self.Size) }
 }
 
 extension swift_metadata_allocation_t: Comparable {
@@ -39,11 +39,11 @@ extension SwiftReflectionContextRef {
 }
 
 extension SwiftReflectionContextRef {
-  typealias ConformanceIterationCallback = (swift_reflection_ptr_t, swift_reflection_ptr_t) -> Void
-  typealias MetadataAllocationIterationCallback = (swift_metadata_allocation_t) -> Void
-  typealias MetadataAllocationBacktraceIterationCallback = (swift_reflection_ptr_t, Int, UnsafePointer<swift_reflection_ptr_t>) -> Void
+  public typealias ConformanceIterationCallback = (swift_reflection_ptr_t, swift_reflection_ptr_t) -> Void
+  public typealias MetadataAllocationIterationCallback = (swift_metadata_allocation_t) -> Void
+  public typealias MetadataAllocationBacktraceIterationCallback = (swift_reflection_ptr_t, Int, UnsafePointer<swift_reflection_ptr_t>) -> Void
 
-  internal var allocations: [swift_metadata_allocation_t] {
+  public var allocations: [swift_metadata_allocation_t] {
     get throws {
       var allocations: [swift_metadata_allocation_t] = []
       try iterateMetadataAllocations {
@@ -53,7 +53,7 @@ extension SwiftReflectionContextRef {
     }
   }
 
-  internal var allocationStacks: [swift_reflection_ptr_t:[swift_reflection_ptr_t]] {
+  public var allocationStacks: [swift_reflection_ptr_t:[swift_reflection_ptr_t]] {
     get throws {
       var stacks: [swift_reflection_ptr_t:[swift_reflection_ptr_t]] = [:]
       try iterateMetadataAllocationBacktraces { allocation, count, stack in
@@ -64,7 +64,7 @@ extension SwiftReflectionContextRef {
     }
   }
 
-  internal func name(type: swift_reflection_ptr_t) -> String? {
+  public func name(type: swift_reflection_ptr_t) -> String? {
     let typeref = swift_reflection_typeRefForMetadata(self, UInt(type))
     if typeref == 0 { return nil }
 
@@ -76,7 +76,7 @@ extension SwiftReflectionContextRef {
     return String(cString: name)
   }
 
-  internal func name(protocol: swift_reflection_ptr_t) -> String? {
+  public func name(protocol: swift_reflection_ptr_t) -> String? {
     guard let name = swift_reflection_copyDemangledNameForProtocolDescriptor(self, `protocol`) else {
       return nil
     }
@@ -85,16 +85,16 @@ extension SwiftReflectionContextRef {
     return String(cString: name)
   }
 
-  internal func name(allocation tag: swift_metadata_allocation_tag_t) -> String?  {
+  public func name(allocation tag: swift_metadata_allocation_tag_t) -> String?  {
     return swift_reflection_metadataAllocationTagName(self, tag).map(String.init)
   }
 
-  internal func isContiguousArray(_ array: swift_reflection_ptr_t) -> Bool {
+  public func isContiguousArray(_ array: swift_reflection_ptr_t) -> Bool {
     guard let name = name(type: array) else { return false }
     return name.hasPrefix("Swift._ContiguousArrayStorage")
   }
 
-  internal func isArrayOfClass(_ array: swift_reflection_ptr_t) -> Bool {
+  public func isArrayOfClass(_ array: swift_reflection_ptr_t) -> Bool {
     guard isContiguousArray(array) else { return false }
 
     let typeref = swift_reflection_typeRefForMetadata(self, UInt(array))
@@ -110,7 +110,7 @@ extension SwiftReflectionContextRef {
     return info.Kind == SWIFT_STRONG_REFERENCE
   }
 
-  internal func arrayCount(_ array: swift_reflection_ptr_t,
+  public func arrayCount(_ array: swift_reflection_ptr_t,
                            _ read: (swift_addr_t, Int) -> UnsafeRawPointer?) -> UInt? {
     // Array layout is: metadata, refCount, count
     let size = MemoryLayout<UInt>.stride * 3
@@ -119,7 +119,7 @@ extension SwiftReflectionContextRef {
     return words[2]
   }
 
-  internal func iterateConformanceCache(_ body: ConformanceIterationCallback) throws {
+  public func iterateConformanceCache(_ body: ConformanceIterationCallback) throws {
     var body = body
     if let error = swift_reflection_iterateConformanceCache(self, {
           let callback = $2!.bindMemory(to: ConformanceIterationCallback.self, capacity: 1)
@@ -129,7 +129,7 @@ extension SwiftReflectionContextRef {
     }
   }
 
-  internal func iterateMetadataAllocations(_ body: MetadataAllocationIterationCallback) throws {
+  public func iterateMetadataAllocations(_ body: MetadataAllocationIterationCallback) throws {
     var body = body
     if let error = swift_reflection_iterateMetadataAllocations(self, {
           let callback = $1!.bindMemory(to: MetadataAllocationIterationCallback.self, capacity: 1)
@@ -139,7 +139,7 @@ extension SwiftReflectionContextRef {
     }
   }
 
-  internal func iterateMetadataAllocationBacktraces(_ body: MetadataAllocationBacktraceIterationCallback) throws {
+  public func iterateMetadataAllocationBacktraces(_ body: MetadataAllocationBacktraceIterationCallback) throws {
     var body = body
     if let error = swift_reflection_iterateMetadataAllocationBacktraces(self, {
           let callback = $3!.bindMemory(to: MetadataAllocationBacktraceIterationCallback.self, capacity: 1)
