+++ tools/swift-inspect/build_script_helper.py
@@ -5,11 +5,13 @@ from __future__ import print_function
 import argparse
 import os
 import subprocess
+import sys
 
 
-def perform_build(args, swiftbuild_path):
-    swiftbuild_args = [
-        swiftbuild_path,
+def perform_swift_action(args, swift_tool_name):
+    swift_tool_path = os.path.join(args.toolchain, "usr", "bin", swift_tool_name)
+    swift_tool_args = [
+        swift_tool_path,
         "--package-path",
         args.package_path,
         "--build-path",
@@ -27,15 +29,21 @@ def perform_build(args, swiftbuild_path):
         "-Xswiftc",
         "-lswiftRemoteMirror"
     ]
+    swift_tool_env = dict(os.environ)
+    swift_tool_env["DYLD_LIBRARY_PATH"] = os.path.join(
+        args.toolchain, 'usr', 'lib', 'swift', 'macosx')
+
     if args.verbose:
-        swiftbuild_args.append("--verbose")
-    print(' '.join(swiftbuild_args))
-    subprocess.call(swiftbuild_args)
+        swift_tool_args.append("--verbose")
+    env_settings = [(k + "=" + v) for k, v in swift_tool_env.items()]
+    print(' '.join(env_settings + swift_tool_args))
+    subprocess.call(swift_tool_args, env=swift_tool_env)
 
 
 def main():
     parser = argparse.ArgumentParser()
     parser.add_argument("--verbose", "-v", action="store_true")
+    parser.add_argument("--action", type=str, required=True)
     parser.add_argument("--package-path", type=str, required=True)
     parser.add_argument("--build-path", type=str, required=True)
     parser.add_argument("--toolchain", type=str, required=True)
@@ -49,8 +57,12 @@ def main():
     if not os.path.isdir(bin_dir):
         os.makedirs(bin_dir)
 
-    swiftbuild_path = os.path.join(args.toolchain, "usr", "bin", "swift-build")
-    perform_build(args, swiftbuild_path)
+    if args.action == "build":
+        perform_swift_action(args, "swift-build")
+    elif args.action == "test":
+        perform_swift_action(args, "swift-test")
+    else:
+        sys.exit("Unknown action " + args.action)
 
 
 if __name__ == "__main__":
